front/refact/RoundedBtn.jsx: moves to src/components/buttons/RoundedBtn.jsx
